Features/file system watcher

See merge request Libraries/LightConversion-Storage-Settings!1

+++ LightConversion.Storage.Settings/Code/Class.ReliableFile.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.IO;
 using System.Text;
 
@@ -14,10 +14,14 @@ namespace LightConversion.Storage.Settings {
         private readonly string _rf1FilePath;
         private readonly string _rf2FilePath;
         private readonly object _lock = new object();
+        private bool _isInitialized;
+        private DateTime _lastWriteDate;
         public string Path { get; }
         public bool AreExceptionsSilent { get; set; }
+        private FileSystemWatcher _fileWatcher;
         public event GeneralEventHandler ErrorOccurred;
         public event GeneralEventHandler InfoReady;
+        public event GeneralEventHandler Changed;
 
         public ReliableFile(string filePath) {
             // Building all the file paths we'll be using in this class.
@@ -26,11 +30,36 @@ namespace LightConversion.Storage.Settings {
             _rf2FilePath = filePath + ".rf2";
         }
 
-        public bool Initialize() {
-            var isOk = false;
+        /// <summary>
+        /// Initialize ReliableFile object. Try to recover file if last write operation failed. Start listening for file changes.
+        /// </summary>
+        /// <exception cref="InvalidOperationException">Thrown when failed to initialize object.</exception>
+        public void Initialize() {
+            if (_isInitialized) return;
+
+            string fileDirectory;
+            string fileName;
+            try {
+                fileName = System.IO.Path.GetFileName(Path);
+            } catch (Exception ex) {
+                throw new InvalidOperationException("Failed to parse file name.", ex);
+            }
+
+            try {
+                fileDirectory = System.IO.Path.GetDirectoryName(System.IO.Path.GetFullPath(Path));
+            } catch (Exception ex) {
+                throw new InvalidOperationException("Failed to parse directory path.", ex);
+            }
+
+            if (string.IsNullOrEmpty(fileDirectory)) {
+                throw new InvalidOperationException("Failed to resolve directory for file:" + Path);
+            }
+
             var mainFileExists = false;
             var rf1FileExists = false;
+
             var rf2FileExists = false;
+
             // Checking what is on the disk.
             lock (_lock) {
                 mainFileExists = File.Exists(Path);
@@ -52,32 +81,25 @@ namespace LightConversion.Storage.Settings {
             // Taking action to fix file, if issues present.
             if (state == FileHealth.Intact) {
                 // All good, file structure is intact.
-                isOk = true;
             } else if (state == FileHealth.Recoverable) {
                 // Something is not right, but rf2 is present, so restoring from it.
-                HandleInfoReady($"Warning. Recovering from \"{_rf2FilePath}\"", "Function Initialize()");
-
+                HandleInfoReady($"Warning. Recovering from \"{_rf2FilePath}\"", $"Function {nameof(Initialize)}()");
                 lock (_lock) {
                     try {
                         File.Delete(Path);
                         File.Delete(_rf1FilePath);
                         File.Move(_rf2FilePath, Path);
-                        isOk = true;
                     } catch (IOException ex) {
-                        HandleNonCriticalError($"File recovery from \"{_rf1FilePath}\" failed because of IOException.", "Function Initialize()", ex);
-                        isOk = false;
+                        HandleNonCriticalError($"File recovery from \"{_rf1FilePath}\" failed because of IOException.", $"Function {nameof(Initialize)}()", ex);
                     }
                 }
             } else if (state == FileHealth.Littered) {
                 // Last write is probably lost, but main file is still there. Just cleaning up.
-                HandleInfoReady($"Warning. Recovering from \"{_rf2FilePath}\". Last write operation is probably lost.", "Function Initialize()");
-
+                HandleInfoReady($"Warning. Recovering from \"{_rf2FilePath}\". Last write operation is probably lost.", $"Function {nameof(Initialize)}()");
                 try {
                     File.Delete(_rf1FilePath);
-                    isOk = true;
                 } catch (IOException ex) {
-                    HandleNonCriticalError($"Deleting temporary \"{_rf1FilePath}\" leftover failed because of IOException.", "Function Initialize()", ex);
-                    isOk = false;
+                    HandleNonCriticalError($"Deleting temporary \"{_rf1FilePath}\" leftover failed because of IOException.", $"Function {nameof(Initialize)}()", ex);
                 }
             } else if (state == FileHealth.Unrecoverable) {
                 // rf2 file is missing, probably saving crashed at some point. rf1 file, if present, is probably corrupt. Can't do much here.
@@ -85,21 +107,44 @@ namespace LightConversion.Storage.Settings {
 
                 try {
                     File.Delete(_rf1FilePath);
-                    isOk = true;
-                } catch (IOException ex) {
-                    HandleNonCriticalError($"Deleting temporary \"{_rf1FilePath}\" leftover failed because of IOException.", "Function Initialize()", ex);
-                    isOk = false;
+                } catch (Exception ex) {
+                    HandleNonCriticalError($"Deleting temporary \"{_rf1FilePath}\" leftover failed because of Exception.", $"Function {nameof(Initialize)}()", ex);
                 }
             }
 
-            return isOk;
+            // Initialize file system watcher.
+            try {
+                _fileWatcher = new FileSystemWatcher(fileDirectory);
+            } catch (Exception ex) {
+                throw new InvalidOperationException("Failed to create FileSystemWatcher.", ex);
+            }
+
+            _fileWatcher.Changed += HandleFileChangedEvent;
+            _fileWatcher.Created += HandleFileCreatedEvent;
+            _fileWatcher.Renamed += HandleFileCreatedEvent;
+            _fileWatcher.Error += OnError;
+            _fileWatcher.Filter = fileName;
+            _fileWatcher.EnableRaisingEvents = true;
+            _isInitialized = true;
+
         }
 
         public bool Exists() {
+            if (_isInitialized == false) {
+                HandleNonCriticalError("Object is not initialized or failed to initialize.", $"Function {nameof(Exists)}()");
+                return false;
+            }
+
             return File.Exists(Path);
         }
 
         public bool TryReadAllText(out string fileContent) {
+            if (_isInitialized == false) {
+                HandleNonCriticalError("Object is not initialized or failed to initialize.", $"Function {nameof(TryReadAllText)}()");
+                fileContent = "";
+                return false;
+            }
+
             var isOk = TryReadAllBytes(out var fileBytes);
 
             if (isOk) fileContent = Encoding.UTF8.GetString(fileBytes);
@@ -109,8 +154,13 @@ namespace LightConversion.Storage.Settings {
         }
 
         public bool TryReadAllBytes(out byte[] fileContent) {
-            bool returnValue;
+            if (_isInitialized == false) {
+                HandleNonCriticalError("Object is not initialized or failed to initialize.", $"Function {nameof(TryReadAllBytes)}()");
+                fileContent = new byte[0];
+                return false;
+            }
 
+            bool returnValue;
             fileContent = new byte[0];
             lock (_lock) {
                 if (File.Exists(Path)) {
@@ -118,11 +168,11 @@ namespace LightConversion.Storage.Settings {
                         fileContent = File.ReadAllBytes(Path);
                         returnValue = true;
                     } catch (IOException ex) {
-                        HandleNonCriticalError("Reading file failed because of IOException.", "Function TryReadAllBytes()", ex);
+                        HandleNonCriticalError("Reading file failed because of IOException.", $"Function {nameof(TryReadAllBytes)}()", ex);
                         returnValue = false;
                     }
                 } else {
-                    HandleNonCriticalError("Reading file failed because it doesn't exist.", "Function TryReadAllBytes()");
+                    HandleNonCriticalError("Reading file failed because it doesn't exist.", $"Function {nameof(TryReadAllBytes)}()");
                     returnValue = false;
                 }
             }
@@ -131,22 +181,43 @@ namespace LightConversion.Storage.Settings {
         }
 
         public bool TryWriteAllText(string textToWrite) {
+            if (_isInitialized == false) {
+                HandleNonCriticalError("Object is not initialized or failed to initialize.", $"Function {nameof(TryWriteAllText)}()");
+                return false;
+            }
+
             return TryWriteAllBytes(Encoding.UTF8.GetBytes(textToWrite));
         }
 
         public bool TryWriteAllBytes(byte[] bytesToWrite) {
+            if (_isInitialized == false) {
+                HandleNonCriticalError("Object is not initialized or failed to initialize.", $"Function {nameof(TryWriteAllBytes)}()");
+                return false;
+            }
+
             bool returnValue;
 
             lock (_lock) {
                 try {
+                    // Let's not rise multiple Changed event with single write operation. Disabling watcher for now.
+                    if (_fileWatcher != null) _fileWatcher.EnableRaisingEvents = false;
+
                     File.WriteAllBytes(_rf1FilePath, bytesToWrite);
                     File.Move(_rf1FilePath, _rf2FilePath);
                     File.Delete(Path);
+
+                    // Enabling watcher before last file operation so we get only one Changed event.
+                    if (_fileWatcher != null) _fileWatcher.EnableRaisingEvents = true;
                     File.Move(_rf2FilePath, Path);
                     returnValue = true;
                 } catch (IOException ex) {
-                    HandleNonCriticalError("Writing to file failed because of IOException.", "Function TryWriteAllBytes()", ex);
+                    HandleNonCriticalError("Writing to file failed because of IOException.", $"Function {nameof(TryWriteAllBytes)}()", ex);
                     returnValue = false;
+                } finally {
+                    // Make sure to reenable watcher before leaving method.
+                    if ((_fileWatcher != null) && (_fileWatcher.EnableRaisingEvents == false)) {
+                        _fileWatcher.EnableRaisingEvents = true;
+                    }
                 }
             }
 
@@ -169,6 +240,36 @@ namespace LightConversion.Storage.Settings {
             }
 
             ErrorOccurred?.Invoke(this, new GeneralEventArgs(fullMessage, source, innerException));
+        }
+
+        private void HandleFileChangedEvent(object sender, FileSystemEventArgs e) {
+            if (Exists() == false) return;
+
+            // When file is edited it rises multiple Changed events. Let's only rise Changed event when file content changes. 
+            var isChanged = true;
+            if (e.ChangeType == WatcherChangeTypes.Changed) {
+                var newLastWriteDate = File.GetLastWriteTime(Path);
+                if (_lastWriteDate == newLastWriteDate) isChanged = false;
+                else _lastWriteDate = newLastWriteDate;
+            }
+
+            if (isChanged) {
+                HandleInfoReady($"{e.ChangeType}: {Path}");
+                Changed?.Invoke(this, new GeneralEventArgs($"Changed: {Path}"));
+            }
+        }
+
+        private void HandleFileCreatedEvent(object sender, FileSystemEventArgs e) {
+            if (Exists() == false) return;
+
+            // It's a new file to watch so last write date is current.
+            _lastWriteDate = File.GetLastWriteTime(Path);
+            HandleInfoReady($"{e.ChangeType}: {Path}");
+            Changed?.Invoke(this, new GeneralEventArgs($"Changed: {Path}"));
+        }
+
+        private void OnError(object sender, ErrorEventArgs e) {
+            HandleNonCriticalError("Error happened in FileSystemWatcher, message:" + e.GetException().Message, $"Function {nameof(OnError)}()");
         }
     }
 }
